Single File decent Algorithm
